aider: The `dirname` command line argument has been modified to accept multiple directory names as a list of strings.
# Context:
USER: make the `dirname` cmd line arg accept many dirnames
ASSISTANT: To make the `dirname` command line argument accept multiple directory names, we need to modify the `main` function in the `benchmark/benchmark.py` file.

We will change the `dirname` argument from a single string to a list of strings. This will allow the user to input multiple directory names separated by spaces.

Here is the change:

```python
benchmark/benchmark.py
<<<<<<< ORIGINAL
def main(
    dirname: str = typer.Argument(..., help="Directory name"),
=======
def main(
    dirname: List[str] = typer.Argument(..., help="Directory names"),
>>>>>>> UPDATED
```

Please note that this change will require modifications in the rest of the code where `dirname` is used, as it will now be a list of strings instead of a single string. We will need to iterate over the list and apply the existing logic to each directory name.

If you agree with this change, I will proceed to identify and modify the rest of the code accordingly.